Tidy up code
